requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
kafka-python==2.0.2